refactor: Split redb service to new crate
